Fix gene numbering
For each decision, there must be a gene to run an instruction.

I was confining movement related instructions to see walkable items, but how does the agent move if they see an enemy?

I added all the "state bools" (seeOther, seeWalkable) to movement handling - so with 3 binary environmental switches, there are 6 movement decisions to make